+++ cmd/validator/main.go
@@ -18,21 +18,21 @@ import (
 	"github.com/setlog/trivrost/pkg/system"
 )
 
-type platformChecks []platformCheck
-
-type platformCheck struct {
-	reason   checkReason
-	platform config.Platform
+type checkDetails struct {
+	reason      checkReason
+	os          string
+	arch        string
+	othersCount int
 }
 
-func (pChecks platformChecks) String() string {
-	if len(pChecks) > 1 {
-		if len(pChecks) > 2 {
-			return fmt.Sprintf("%s on platform %v and %d others", pChecks[0].reason, pChecks[0].platform, len(pChecks)-1)
+func (cd checkDetails) String() string {
+	if cd.othersCount > 0 {
+		if cd.othersCount > 1 {
+			return fmt.Sprintf("%s on platform %s-%s and %d others", cd.reason, cd.os, cd.arch, cd.othersCount)
 		}
-		return fmt.Sprintf("%s on platform %v and one other", pChecks[0].reason, pChecks[0].platform)
+		return fmt.Sprintf("%s on platform %s-%s and one other", cd.reason, cd.os, cd.arch)
 	}
-	return fmt.Sprintf("%s on platform %v", pChecks[0].reason, pChecks[0].platform)
+	return fmt.Sprintf("%s on platform %s-%s", cd.reason, cd.os, cd.arch)
 }
 
 type checkReason int
@@ -69,167 +69,102 @@ func main() {
 	}
 
 	if !skipurlcheck {
-		urlMap, collectErr := collectURLs(data, skipJarCheck)
-		urlErr := checkURLs(urlMap, skipJarCheck)
-		cmdErr := checkCommands(data)
-		if collectErr != nil {
-			printError(collectErr)
-		}
-		if urlErr != nil {
-			printError(urlErr)
-		}
-		if cmdErr != nil {
-			printError(cmdErr)
-		}
-		if collectErr != nil || urlErr != nil || cmdErr != nil {
-			fatalf("There were errors which need to be fixed.")
-		}
+		checkURLs(data, filePath, skipJarCheck)
 	}
 }
 
-func checkCommands(data []byte) error {
-	failCount := 0
-	for _, os := range []string{"windows", "darwin", "linux"} {
-		for _, arch := range []string{"386", "amd64"} {
-			deploymentConfig := config.ParseDeploymentConfig(strings.NewReader(string(data)), os, arch)
-			for _, command := range deploymentConfig.Execution.Commands {
-				if !isAbsForOS(command.Name, os) {
-					if !willDownloadCommandFile(command, deploymentConfig) {
-						printError(fmt.Errorf("Command \"%s\" would not be available on platform %s-%s: missing bundle entry", command.Name, os, arch))
-						failCount++
-					} else {
-						fmt.Printf("OK: Binary for command \"%s\" would be downloaded on platform %s-%s (which requires it).\n", command.Name, os, arch)
-					}
-				}
-			}
-		}
-	}
-	if failCount > 0 {
-		return fmt.Errorf("%d cases of commands' bundles not being configured to be downloaded", failCount)
-	}
-	return nil
-}
+func checkURLs(data []byte, filePath string, skipJarCheck bool) {
+	urlMap, success := collectURLs(data, skipJarCheck)
 
-func willDownloadCommandFile(command config.Command, deploymentConfig *config.DeploymentConfig) bool {
-	commandNameUnix := strings.ReplaceAll(command.Name, `\`, "/")
-	bundleName := misc.FirstElementOfPath(commandNameUnix)
-	return getBundleURL(bundleName, deploymentConfig) != ""
-}
-
-func checkURLs(urlMap map[string]platformChecks, skipJarCheck bool) error {
 	waitgroup := sync.WaitGroup{}
 	waitgroup.Add(len(urlMap))
 	var gotCertError bool
 
 	// check all url in parallel
 	var errorCount int32
-	for url, pChecks := range urlMap {
-		go func(url string, pChecks platformChecks) {
+	for url, details := range urlMap {
+		go func(url string, details checkDetails) {
 			defer waitgroup.Done()
 			code, err := getUrlHeadResult(url)
 			if err != nil {
 				_, isUnknownAuthorityError := err.(x509.UnknownAuthorityError)
 				gotCertError = gotCertError || isUnknownAuthorityError
-				fmt.Printf("\033[0;91mHTTP HEAD request to URL '%s' failed: %v. (Check reason: %v)\033[0m\n", url, err, pChecks[0])
+				fmt.Printf("\033[0;91mHTTP HEAD request to URL '%s' failed: %v. (Check reason: %v)\033[0m\n", url, err, details)
 				atomic.AddInt32(&errorCount, 1)
 			} else if code != http.StatusOK {
-				fmt.Printf("\033[0;91mHTTP HEAD request to URL '%s' yielded bad response code %d. (Check reason: %v).\033[0m\n", url, code, pChecks[0])
+				fmt.Printf("\033[0;91mHTTP HEAD request to URL '%s' yielded bad response code %d. (Check reason: %v)\033[0m\n", url, code, details)
 				atomic.AddInt32(&errorCount, 1)
 			} else {
-				fmt.Printf("OK: Resource %s is available. (Reason for check: %v)\n", url, pChecks[0])
+				fmt.Printf("OK: Resource %s is available. (Reason for check: %v)\n", url, details)
 			}
-		}(url, pChecks)
+		}(url, details)
 	}
 	waitgroup.Wait()
 	if errorCount > 0 {
 		if gotCertError {
 			fmt.Printf("\033[0;91mThere was at least one certificate-related error. The system's certificate pool may be out of date.\033[0m\n")
 		}
-		return fmt.Errorf("%d out of %d tested URLs from the deployment-config do not point to valid resources", errorCount, len(urlMap))
+		fatalf("%d out of %d tested URLs from \"%s\" do not point to valid resources.", errorCount, len(urlMap), filePath)
+		success = false
 	}
-	return nil
+	if !success {
+		os.Exit(1)
+	}
 }
 
-func collectURLs(data []byte, skipJarCheck bool) (urlMap map[string]platformChecks, err error) {
-	urlMap = make(map[string]platformChecks)
-	failCount := 0
-	for _, os := range []string{"windows", "darwin", "linux"} {
+func collectURLs(data []byte, skipJarCheck bool) (urlMap map[string]checkDetails, success bool) {
+	urlMap = make(map[string]checkDetails)
+	success = true
+	for _, operatingsystem := range []string{"windows", "darwin", "linux"} {
 		for _, arch := range []string{"386", "amd64"} {
-			deploymentConfig := config.ParseDeploymentConfig(strings.NewReader(string(data)), os, arch)
+			deploymentConfig := config.ParseDeploymentConfig(strings.NewReader(string(data)), operatingsystem, arch)
 			for _, update := range deploymentConfig.LauncherUpdate {
-				addUrlWithDetails(urlMap, update.BundleInfoURL, platformCheck{reasonUpdate, config.Platform{OS: os, Arch: arch}})
+				addUrlWithDetails(urlMap, update.BundleInfoURL, checkDetails{reasonUpdate, operatingsystem, arch, 0})
 			}
 			for _, update := range deploymentConfig.Bundles {
-				addUrlWithDetails(urlMap, update.BundleInfoURL, platformCheck{reasonBundle, config.Platform{OS: os, Arch: arch}})
+				addUrlWithDetails(urlMap, update.BundleInfoURL, checkDetails{reasonBundle, operatingsystem, arch, 0})
 			}
 			for _, command := range deploymentConfig.Execution.Commands {
-				if err := collectCommandURLs(urlMap, deploymentConfig, os, arch, command, skipJarCheck); err != nil {
-					failCount++
-					printError(err)
-				}
+				success = collectCommandURLs(urlMap, deploymentConfig, operatingsystem, arch, command, skipJarCheck) && success
 			}
 		}
 	}
-	if failCount > 0 {
-		err = fmt.Errorf("%d commands were malformed", failCount)
-	}
-	return urlMap, err
+	return urlMap, success
 }
 
-func addUrlWithDetails(urlMap map[string]platformChecks, url string, pCheck platformCheck) {
-	presentChecks, ok := urlMap[url]
+func addUrlWithDetails(urlMap map[string]checkDetails, url string, details checkDetails) {
+	presentDetails, ok := urlMap[url]
 	if ok {
-		urlMap[url] = append(presentChecks, pCheck)
+		presentDetails.othersCount++
+		urlMap[url] = presentDetails
 	} else {
-		urlMap[url] = platformChecks{pCheck}
+		urlMap[url] = details
 	}
 }
 
-func isAbsForOS(filePath string, os string) bool {
-	if os == system.OsWindows {
-		return filePath != "" && ((filePath[0] >= 'A' && filePath[0] <= 'Z') || (filePath[0] >= 'a' && filePath[0] <= 'z')) &&
-			(len(filePath) == 1 || (filePath[1] == ':' && (len(filePath) == 2 || filePath[2] == '\\' || filePath[2] == '/')))
-	}
-	return path.IsAbs(filePath)
-}
-
-func collectCommandURLs(urlMap map[string]platformChecks, deploymentConfig *config.DeploymentConfig, os, arch string, command config.Command, skipJarCheck bool) error {
-	checks, err := getCommandURLs(deploymentConfig, os, arch, command, skipJarCheck)
-	if err != nil {
-		return err
-	}
-	for k, v := range checks {
-		urlMap[k] = v
-	}
-	return nil
-}
-
-func getCommandURLs(deploymentConfig *config.DeploymentConfig, os, arch string, command config.Command, skipJarCheck bool) (map[string]platformChecks, error) {
+func collectCommandURLs(urlMap map[string]checkDetails, deploymentConfig *config.DeploymentConfig, os, arch string, command config.Command, skipJarCheck bool) (success bool) {
 	commandNameUnix := strings.ReplaceAll(command.Name, `\`, "/")
-	bundleName := misc.FirstElementOfPath(commandNameUnix)
-	if isAbsForOS(command.Name, os) {
-		fmt.Printf("\033[0;96mNote: cannot validate absolute command path for bundle \"%s\". (Command \"%s\" for platform %s-%s is absolute path).\033[0m\n", bundleName, command.Name, os, arch)
-		return nil, nil
-	} else if !strings.Contains(commandNameUnix, "/") {
-		return nil, fmt.Errorf("Could not get bundle URL for bundle \"%s\": relative command \"%s\" for platform %s-%s does not descend into a bundle directory", bundleName, command.Name, os, arch)
+	if path.IsAbs(commandNameUnix) || !strings.Contains(commandNameUnix, "/") {
+		return
 	}
+	bundleName := misc.FirstElementOfPath(commandNameUnix)
 	bundleURL := getBundleURL(bundleName, deploymentConfig)
 	if bundleURL == "" {
-		return nil, fmt.Errorf("No BaseURL configured or inferable for bundle \"%s\": not set. (Required for command \"%s\" on platform %s-%s)", bundleName, command.Name, os, arch)
+		fmt.Printf("\033[0;91mCould not get bundle URL for bundle \"%s\" for platform %s-%s. (Required for command \"%s\")\033[0m\n", bundleName, os, arch, command.Name)
+		return false
 	}
 	binaryURL := misc.MustJoinURL(bundleURL, stripFirstPathElement(commandNameUnix))
 	if os == system.OsWindows && !strings.HasSuffix(binaryURL, ".exe") {
 		binaryURL += ".exe"
 	}
-	urlMap := make(map[string]platformChecks)
-	addUrlWithDetails(urlMap, binaryURL, platformCheck{reasonCommand, config.Platform{OS: os, Arch: arch}})
+	addUrlWithDetails(urlMap, binaryURL, checkDetails{reasonCommand, os, arch, 0})
 	if !skipJarCheck {
 		if strings.HasSuffix(binaryURL, "/java.exe") || strings.HasSuffix(binaryURL, "/javaw.exe") ||
 			strings.HasSuffix(binaryURL, "/java") {
 			collectJarURL(urlMap, deploymentConfig, command, os, arch)
 		}
 	}
-	return urlMap, nil
+	return true
 }
 
 func stripFirstPathElement(s string) string {
@@ -241,7 +176,7 @@ func stripFirstPathElement(s string) string {
 	return strings.Join(parts[1:], "/")
 }
 
-func collectJarURL(urlMap map[string]platformChecks, deploymentConfig *config.DeploymentConfig, command config.Command, os, arch string) {
+func collectJarURL(urlMap map[string]checkDetails, deploymentConfig *config.DeploymentConfig, command config.Command, os, arch string) {
 	check := false
 	for _, arg := range command.Arguments {
 		if check {
@@ -249,7 +184,7 @@ func collectJarURL(urlMap map[string]platformChecks, deploymentConfig *config.De
 			bundleName := misc.FirstElementOfPath(jarPath)
 			bundleURL := getBundleURL(bundleName, deploymentConfig)
 			jarURL := misc.MustJoinURL(bundleURL, stripFirstPathElement(jarPath))
-			addUrlWithDetails(urlMap, jarURL, platformCheck{reasonJar, config.Platform{OS: os, Arch: arch}})
+			addUrlWithDetails(urlMap, jarURL, checkDetails{reasonJar, os, arch, 0})
 			break
 		}
 		if arg == "-jar" {
@@ -295,16 +230,7 @@ func parseFlags() (string, bool, bool) {
 	return deploymentConfigPath, *skipurlcheck, *skipJarCheck
 }
 
-func fatalf(msg string, args ...interface{}) {
-	fmt.Printf("\033[0;91mFatal: "+msg+".\033[0m\n", args...)
+func fatalf(formatMessage string, args ...interface{}) {
+	fmt.Printf("\033[0;91mFatal: "+formatMessage+"\033[0m\n", args...)
 	os.Exit(1)
-}
-
-func fatalErr(err error) {
-	fmt.Printf("\033[0;91mFatal: %v.\033[0m\n")
-	os.Exit(1)
-}
-
-func printError(err error) {
-	fmt.Printf("\033[0;91m%v.\033[0m\n")
 }